feat: reorganize project display and enhance AppHeader navigation
